backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymysql==1.1.1
python-dotenv==1.0.1
cryptography==42.0.8
PyJWT==2.8.0
bcrypt==4.1.3
PyJWT==2.8.0
werkzeug==3.0.3
